--- src/assets/css/Question.css
@@ -1,43 +1,0 @@
-h2 {
-  font-weight: bold;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-}
-
-fieldset {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-}
-
-label {
-  margin-bottom: 0;
-}
-
-.center {
-  display: flex;
-  flex-direction: column;
-}
-
-.answer {
-  display: flex;
-  justify-content: left;
-  margin-bottom: 0.5em;
-  font-size: 1.5em;
-}
-
-.labelAnswer {
-  margin-left: 0.3em;
-}
-
-input.submit {
-  background-color: black;
-  color: white;
-  border-style: none;
-  border-radius: 1em;
-  padding: 1em 0;
-  margin-bottom: 0.5em;
-  font-size: 1.1em;
-}
