+++ src/components/Home.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+function Home () {
+return(
+  <div>
+  </div>
+)}
+export default Home
